fix: collateral fixes
