+++ src/app/loans/loans-account-stepper/loans-account-charges-step/loans-account-charges-step.component.html
@@ -116,57 +116,6 @@
     <tr mat-row *matRowDef="let row; columns: overdueChargesDisplayedColumns;"></tr>
   </table>
 
-  <mat-divider fxFlex="98%"></mat-divider>
-
-  <div fxFlex="50%">
-    <div fxLayout="column" fxFlex="50%" class="tableName">
-      <h4 fxFlex="98%" class="mat-h4">Collaterals Data</h4>
-    </div>
-
-    <div fxLayout="column" fxFlex="50%">
-      <div fxLayout="row" fxLayoutAlign="flex-end">
-        <button mat-raised-button color="primary" (click)="addCollateral()">
-          <fa-icon icon="plus"></fa-icon>&nbsp;&nbsp;Add
-        </button>
-      </div>
-    </div>
-  </div>
-
-  <table mat-table class="mat-elevation-z1" [dataSource]="collateralDataSource">
-
-    <ng-container matColumnDef="type">
-      <th mat-header-cell *matHeaderCellDef> Name </th>
-      <td mat-cell *matCellDef="let ele"> {{ ele.type.name }} </td>
-    </ng-container>
-
-    <ng-container matColumnDef="value">
-      <th mat-header-cell *matHeaderCellDef> Quantity </th>
-      <td mat-cell *matCellDef="let ele"> {{ ele.value }} </td>
-    </ng-container>
-
-    <ng-container matColumnDef="totalValue">
-      <th mat-header-cell *matHeaderCellDef> Total Value </th>
-      <td mat-cell *matCellDef="let ele"> {{ ele.type.basePrice*ele.value }} </td>
-    </ng-container>
-    
-    <ng-container matColumnDef="totalCollateralValue">
-      <th mat-header-cell *matHeaderCellDef> Total Collateral Value </th>
-      <td mat-cell *matCellDef="let ele"> {{ ele.type.pctToBase*ele.type.basePrice*ele.value/100 }} </td>
-    </ng-container>
-
-    <ng-container matColumnDef="action">
-      <th mat-header-cell *matHeaderCellDef> Actions </th>
-      <td mat-cell *matCellDef="let ele; let collateralIndex = index">
-        <button mat-icon-button color="warn" (click)="deleteCollateral(collateralIndex)">
-          <fa-icon icon="trash"></fa-icon>
-        </button>
-      </td>
-    </ng-container>
-
-    <tr mat-header-row *matHeaderRowDef="loanCollateralDisplayedColumns"></tr>
-    <tr mat-row *matRowDef="let row; columns: loanCollateralDisplayedColumns;"></tr>
-  </table>
-
 </div>
 
 <div fxLayout="row" class="margin-t" fxLayout.xs="column" fxLayoutAlign="center" fxLayoutGap="2%">
