+++ src/app/loans/loans-account-stepper/loans-account-charges-step/loans-account-charges-step.component.ts
@@ -31,10 +31,6 @@ export class LoansAccountChargesStepComponent implements OnInit, OnChanges {
   @Input() loansAccountTemplate: any;
   // @Input() loansAccountFormValid: LoansAccountFormValid
   @Input() loansAccountFormValid: boolean;
-  // @Input collateralOptions: Collateral Options
-  @Input() collateralOptions: any;
-  // @Input loanPrincipal: Loan Principle
-  @Input() loanPrincipal: any;
 
   /** Charges Data */
   chargeData: any;
@@ -48,8 +44,6 @@ export class LoansAccountChargesStepComponent implements OnInit, OnChanges {
   chargesDisplayedColumns: string[] = ['name', 'chargeCalculationType', 'amount', 'chargeTimeType', 'date', 'action'];
   /** Columns to be displayed in overdue charges table. */
   overdueChargesDisplayedColumns: string[] = ['name', 'type', 'amount', 'collectedon'];
-  /** Columns to be displayed in collateral table. */
-  loanCollateralDisplayedColumns: string[] = ['type', 'value', 'totalValue', 'totalCollateralValue', 'action'];
   /** Component is pristine if there has been no changes by user interaction */
   pristine = true;
   /** Check if value of collateral added  is more than principal amount */
@@ -213,56 +207,6 @@ export class LoansAccountChargesStepComponent implements OnInit, OnChanges {
     });
   }
 
-  /**
-   * Add a Collateral to the loan
-   */
-  addCollateral() {
-    const addCollateralDialogRef = this.dialog.open(LoansAccountAddCollateralDialogComponent, {
-      data: { collateralOptions: this.collateralOptions }
-    });
-    console.log(this.collateralOptions);
-    addCollateralDialogRef.afterClosed().subscribe((response: any) => {
-      console.log(this.loanPrincipal);
-      if (response.data) {
-        const collateralData = {
-          type: response.data.value.collateral,
-          value: response.data.value.quantity,
-        };
-
-        this.totalCollateralValue += collateralData.type.pctToBase * collateralData.type.basePrice * collateralData.value / 100;
-        this.collateralDataSource = this.collateralDataSource.concat(collateralData);
-        this.collateralOptions = this.collateralOptions.filter((user: any) => user.collateralId !== response.data.value.collateral.collateralId);
-        if (this.loanPrincipal < this.totalCollateralValue) {
-          this.isCollateralSufficient = true;
-        } else {
-          this.isCollateralSufficient = false;
-        }
-      }
-    });
-  }
-  /**
-   * Delete a added collateral from loan
-   * @param id ID od the collateral to be deleted
-   */
-  deleteCollateral(id: any) {
-    const deleteCollateralDialogRef = this.dialog.open(DeleteDialogComponent, {
-      data: { deleteContext: `collateral` }
-    });
-    deleteCollateralDialogRef.afterClosed().subscribe((response: any) => {
-      if (response.delete) {
-        const removed: any = this.collateralDataSource.splice(id, 1);
-        this.collateralOptions = this.collateralOptions.concat(removed[0].type);
-        this.totalCollateralValue -= removed[0].type.pctToBase * removed[0].type.basePrice * removed[0].value / 100;
-        this.collateralDataSource = this.collateralDataSource.concat([]);
-        this.pristine = false;
-        if (this.loanPrincipal < this.totalCollateralValue) {
-          this.isCollateralSufficient = true;
-        } else {
-          this.isCollateralSufficient = false;
-        }
-      }
-    });
-  }
 
   /**
    * Returns Loans Account Charges and Collateral Form
@@ -270,7 +214,6 @@ export class LoansAccountChargesStepComponent implements OnInit, OnChanges {
   get loansAccountCharges() {
     return {
       charges: this.chargesDataSource,
-      collateral: this.collateralDataSource
     };
   }
